refactor: controller가 가지고 있는 cars 필드 제거
