+++ src/main/java/racing/controller/ConsoleRacingGameController.java
@@ -1,6 +1,7 @@
 package racing.controller;
 
-import racing.RandomNumberGenerator;
+import racing.ui.input.InputView;
+import racing.util.RandomNumberGenerator;
 import racing.domain.Car;
 import racing.domain.Cars;
 import racing.service.RacingGameService;
@@ -8,21 +9,22 @@ import racing.ui.output.OutputView;
 
 public class ConsoleRacingGameController {
 
-    private final Cars cars;
     private final RacingGameService racingGameService;
 
-    public ConsoleRacingGameController(Cars cars) {
-        this.cars = cars;
+    public ConsoleRacingGameController() {
         this.racingGameService = new RacingGameService();
     }
 
-    public void start(int count) {
+    public void start() {
         OutputView.printPhrase();
-        move(count);
+        Cars cars = racingGameService.createCars(InputView.inputCarsName());
+        int count = Integer.parseInt(InputView.inputCount());
+
+        moveCar(cars, count);
         OutputView.printResult(cars);
     }
 
-    private void move(int count) {
+    private void moveCar(Cars cars, int count) {
         RandomNumberGenerator generator = new RandomNumberGenerator();
         for (int i = 0; i < count; i++) {
             for (Car car : cars.getCars()) {
