+++ src/main/java/racing/controller/WebRacingGameController.java
@@ -4,7 +4,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import racing.RandomNumberGenerator;
+import racing.util.RandomNumberGenerator;
 import racing.controller.dto.request.RacingGameInfoRequest;
 import racing.controller.dto.response.RacingGameResultResponse;
 import racing.domain.Car;
