+++ src/test/java/racing/RandomNumberGeneratorTest.java
@@ -3,6 +3,7 @@ package racing;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.util.RandomNumberGenerator;
 
 class RandomNumberGeneratorTest {
